fix: applied feedback on moving helper funcs
